rust/SDL2 is being a pain again

+++ src/app_selector.rs
@@ -77,7 +77,7 @@ impl Component for AppSelector {
             app.camera.push();
 
             // app.camera.set_ipos(50, 0);
-            app.camera.translate(cam_pos.0 as f64, cam_pos.1 as f64, 0f64);
+            app.camera.translate(cam_pos.0 as f32, cam_pos.1 as f32, 0f32);
             app.camera.set_viewport((45, 0, app.window_size.0 - 45, app.window_size.1));
 
             self.game_app.update(app);
@@ -96,7 +96,7 @@ impl Component for AppSelector {
 
             // app.camera.set_ipos(50, 0);
 
-            app.camera.translate(cam_pos.0 as f64, cam_pos.1 as f64, 0f64);
+            app.camera.translate(cam_pos.0 as f32, cam_pos.1 as f32, 0f32);
 
             app.camera.set_viewport((50, 0, app.window_size.0 - 50, app.window_size.1));
             
@@ -119,7 +119,7 @@ impl Component for AppSelector {
 
             // app.camera.set_ipos(50, 0);
 
-            app.camera.translate(cam_pos.0 as f64, cam_pos.1 as f64, 0f64);
+            app.camera.translate(cam_pos.0 as f32, cam_pos.1 as f32, 0f32);
 
             app.camera.set_viewport((45, 0, app.window_size.0 - 45, app.window_size.1));
