+++ src/button.rs
@@ -41,7 +41,7 @@ impl Component for Button {
         app.camera.push();
 
         // app.camera.set_position(self.position.0 as f32 / app.window_size.1 as f32, self.position.1 as f32 / app.window_size.1 as f32);
-        app.camera.translate(self.position.0 as f64 / app.window_size.1 as f64, self.position.1 as f64 / app.window_size.1 as f64, 0f64);
+        app.camera.translate(self.position.0 as f32 / app.window_size.1 as f32, self.position.1 as f32 / app.window_size.1 as f32, 0f32);
 
         for child in &mut self.children {
             child.update(app);
